--- openlibrary/macros/FulltextSearchBox.html
@@ -1,57 +1,0 @@
-$def with(query, results, page=1)
-
-
-$if results and results.get('hits'):
-  $ hits = results['hits'].get('hits', [])
-  $ num_found = results['hits'].get('total', 0)
-  $ ia_base_url = "https://archive.org"
-  <div id="searchResults">
-    <div>Browse $num_found Search Inside Matches</div>
-    <ul class="list-books">
-      $for doc in hits:
-        $ ia = doc.get('fields', {}).get('identifier', [''])[0]
-        $ snippet = doc.get('highlight', {}).get('text', [''])[0]
-        $ page_nums = doc.get('fields', {}).get('page_num', [])
-        $ page = ', '.join([str(num) for num in page_nums])
-        $ title = doc.get('fields', {}).get('meta_title', [''])[0]
-        $ author_info = doc.get('fields', {}).get('meta_creator', [''])[0].split(', ')
-        $ author_name = ''
-        $if author_info and author_info[0]:
-          $if len(author_info) == 1:
-            $ author_name = author_info[0]
-          $elif len(author_info) > 1:
-            $ author_name = author_info[1] + ' ' + author_info[0]
-
-        $if snippet:
-          <section class="fulltext-excerpts searchResultItem">
-            <div class="sri__main">
-              <span class="bookcover">
-                $ cover = "%s/services/img/%s" % (ia_base_url, ia)
-                <a href="$ia_base_url/details/$ia">
-                  <img itemprop="image"
-                    src="$cover"
-                  />
-                </a>
-            </span>
-            <div class="details">
-              <div class="resultTitle">
-                <h3 class="booktitle">
-                  <a href="/books/$ia/$title" class="results">$title</a>
-                </h3>
-              </div>
-            </div>
-            </div>
-            <span class="bookauthor">
-              <div>$author_info $author_name</div>
-            </span>
-            <div class="serp-extra">
-              <ul>
-                  $if snippet:
-                    <li class="fulltext-excerpt">
-                        &hellip;$:(snippet.replace("<", "&laquo;").replace(">", "&raquo;").replace("{{{", "<mark class='highlight'><strong>").replace("}}}", "</strong></mark>"))&hellip;
-                    </li>
-              </ul>
-            </div>
-          </section>
-      </ul>
-  </div>
