+++ openlibrary/plugins/inside/code.py
@@ -20,6 +20,7 @@ class search_inside(delegate.page):
         query = i.q
         page = int(i.page)
         results = fulltext_search(query, page=page, limit=RESULTS_PER_PAGE)
+        print('*************RESULTS FROM FULLTEXT****************', results)
         search_time = time() - search_start
 
         return render_template(
