+++ openlibrary/templates/work_search.html
@@ -119,7 +119,7 @@ $ )
         $# Temporarily (2020-03-26) disable automatically showing full-text
         $# results because of performance issues due to increased load. See
         $# this commit to revert.
-        <div id="fulltext-search-box" data-query="$query">TESTING</div>
+        <div id="fulltext-search-suggestion" data-query="$query">TESTING</div>
         <!-- <center>
           <div class="red">$_("No results found.")</div>
           <hr>
